Fix warning (implicit conversion double to std::size_t)

+++ Point_set_processing_3/test/Point_set_processing_3/wlop_simplify_and_regularize_test.cpp
@@ -56,7 +56,7 @@ void test_wlop_simplify_and_regularize(
 
   // Make room for sample points
   std::vector<Point> points_sampled;
-  points_sampled.resize(points.size() * (retain_percentage / 100.));
+  points_sampled.resize(static_cast<std::size_t>(points.size() * (retain_percentage / 100.)));
 
   output.clear();
   // Run algorithm
